+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Released on Tuesday, May 31 2022.
 - Updated to use AngleSharp 0.17
 - Fixed casing issue with color, timing, and gradient functions (#109)
 - Fixed shorthand properties using `inherit` being omitted (#100)
+- Added support for `conic-gradient` (#101)
 
 # 0.16.4